Querify object_safety_violations.
