+++ src/librustc_typeck/check/wfcheck.rs
@@ -223,7 +223,7 @@ fn check_object_unsafe_self_trait_by_name(tcx: TyCtxt<'_>, item: &hir::TraitItem
         _ => {}
     }
     if !trait_should_be_self.is_empty() {
-        if rustc_infer::traits::object_safety_violations(tcx, trait_def_id).is_empty() {
+        if tcx.object_safety_violations(trait_def_id).is_empty() {
             return;
         }
         let sugg = trait_should_be_self.iter().map(|span| (*span, "Self".to_string())).collect();
